% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeness.R
\name{completeness}
\alias{completeness}
\title{SMR01 completeness}
\usage{
completeness(quarter = c("previous", "current"), level = c("board",
  "scotland"), first_day)
}
\arguments{
\item{quarter}{Should completeness data be returned for the current quarter,
or the previous one?}

\item{level}{Should completeness data be returned at board or Scotland level?}

\item{first_day}{The first day of the latest quarter. Must be the first day
of either January, April, July or October. Only one value of \code{first_day}
is valid at any one time. When the spreadsheet is updated with a new
quarter's data, the previously valid value of \code{first_day} will now
return an error, and a new value (the previous value + 3 months) should be
supplied instead.}
}
\value{
When \code{level} is set equal to \code{board}, \code{completeness}
returns a sentence listing the boards which had an SMR01 completeness level
of less than 95 percent for the specified \code{quarter}. When \code{level}
is set equal to \code{scotland}, \code{completeness} returns Scotland's SMR01
completeness level for the specified \code{quarter} as a percentage.
}
\description{
\code{completeness} takes the first day of the latest quarter
of data which are included in the current HSMR publication. It extracts
completeness information for the SMR01 dataset from the
\href{https://www.isdscotland.org/Products-and-Services/Data-Support-and-Monitoring/SMR-Completeness/_docs/SMR_Estimates.xlsx}{SMR Completeness Estimates spreadsheet}
published on the
\href{https://www.isdscotland.org/}{ISD website}. It extracts the
completeness percentages for the two most recent quarters for all health
boards and Scotland.
}
\details{
The \code{completeness} function assumes that the list of health
board names (including Scotland) is contained in cells B31:B47 in the
spreadsheet. Should this change, the function should be edited accordingly.

The URL of the spreadsheet is occasionally modified. Should this happen, the
function should again be edited accordingly.
}
