% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_trends.R
\name{create_trends}
\alias{create_trends}
\title{minimal tidy data set for long term trends in crude mortality report
production.}
\usage{
create_trends(SMR01, GRO, pop, dep)
}
\arguments{
\item{SMR01}{Input tibble for admissions, see details.}

\item{GRO}{Input tibble for deaths, see details.}

\item{pop}{Input tibble for population estimates.}

\item{dep}{Input tibble for deprivation lookup.}
}
\value{
If the class is not instantiated correctly, nothing is returned.
}
\description{
Returns the minimal tidy data set for long terms trend data
which feeds into the reproducible analytical pipeline of the quarterly
Hospital Standardised Mortality Ratios publicatoin.
}
\details{
\code{create_trends} expects a \code{tibble} with data extracted
from SMR01, a \code{tibble} with data extracted from NRS deaths records, a
\code{tibble} with population estimates and a \code{tibble} with a
lookup file for deprivation.
}
\examples{


}
