% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnel_text.R
\name{funnel_text}
\alias{funnel_text}
\title{Create text for Funnel Plot outliers}
\usage{
funnel_text(smr_data)
}
\arguments{
\item{smr_data}{Should the final\code{tibble} output from the \code{smr_data}
function.}
}
\description{
The publication summary documents for the HSMR publication has main points
relating to outliers on the funnel plot. The \code{funnel_text} function
identifies outlying hospitals and returns the text for the publication
document.
}
\examples{
funnel_text(smr_data)

}
