% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsmr_hid.R
\name{hsmr_hid}
\alias{hsmr_hid}
\title{Hospital Intelligence Dashboard.}
\usage{
hsmr_hid(smr_data, trend_data, end_date)
}
\arguments{
\item{data}{The input \code{tibble} which is the output from \code{smr_data}
or \code{create_trends}.}
}
\value{
If the class is not initiated correctly, nothing is returned.
}
\description{
\code{hsmr_hid} reformats the output from \code{smr_data}
and \code{trend_data} in order to fit the template agreed for HID.
It adds variables \code{scot_ratio}, \code{Topic}, \code{Indicator}
and \code{Frequency}; and reformats \code{location}, \code{hb}.
Dropping all other variables except the following which are renamed:
\code{hb}, \code{location_name}, \code{deaths}, \code{pred}, \code{pats},
\code{smr}, \code{crd_rate}, \code{death_scot}, \code{scot_deaths},
\code{pred_scot}, \code{scot_pats} and \code{st_err}.
}
\details{
\href{https://www.isdscotland.org/Health-Topics/Quality-Indicators/
Hospital-Scorecard//}{Hospital Intelligence Dashboard} is a management
information system to be interpreted and used nationally by the Scottish
Government and locally by NHS Boards and clinicians. It allows comparison
of each NHS Boards' performance against the national average, and provides
trend analysis to help monitor variation over time for each indicator where
appropriate. The data are presented by NHS Board and hospital of treatment.
}
