% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mit_available.R
\name{mit_available}
\alias{mit_available}
\title{Management Information Tool made available to health boards}
\usage{
mit_available(end_date)
}
\arguments{
\item{end_date}{The cut-off date for data to be included in the current HSMR
publication, supplied with \code{Date} class. Must be the final day of
either March, June, September or December.}
}
\description{
\code{mit_available} takes the final date for which SMR data
are included in the current publication. It uses this cut-off date to
calculate the date on which data for the current publication are made
available to select health board employees via the management information
tool.
}
\details{
The management information tool is a dashboard which accompanies
the full report, summary report and Excel tables as part of the HSMR
publication. It contains all of the data which the other parts of the
publication are based upon.

The management tool is made available to health boards on the last Friday of
the month, the month before publication.
}
\examples{
mit_available(end_date = lubridate::dmy(30092018))
format(mit_available(end_date = lubridate::dmy(31122018)), "\%d \%B \%Y")
format(mit_available(end_date = lubridate::dmy(31032019)), "\%d/\%m/\%Y")

}
