% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtr.R
\name{qtr}
\alias{qtr}
\alias{qtr_end}
\alias{qtr_prev}
\title{Assign a date to a quarter}
\usage{
qtr(first_day, format = c("long", "short"))

qtr_end(first_day, quarter = c("current", "next"))

qtr_prev(first_day)
}
\arguments{
\item{first_day}{The first day of a quarter, supplied with \code{Date} class.}

\item{format}{Should the quarter be provided in short or long format?}

\item{quarter}{Should the final month of the current or next quarter be
provided?}
}
\description{
The qtr functions take the first day of a quarter (the first of
either January, April, July or October) and calculate the relevant
quarter-related value from it.

\itemize{
\item \code{qtr} returns the current quarter in either short or long format.

\item \code{qtr_end} returns the last month in the quarter, either of the
current or following quarter.

\item \code{qtr_prev} returns the previous quarter in long format.
}
}
\examples{
qtr(first_day = lubridate::dmy(01012018), format = "short")
qtr(first_day = lubridate::dmy(01102018), format = "long")

qtr_end(first_day = lubridate::dmy(01072018), quarter = "current")
qtr_end(first_day = lubridate::dmy(01042018), quarter = "next")

qtr_prev(first_day = lubridate::dmy(01012018))
}
