% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smr_pmorbs.R
\name{smr_pmorbs}
\alias{smr_pmorbs}
\title{Prior-morbidities Wrangling for HSMR data}
\usage{
smr_pmorbs(smr01, smr01_minus5, morbs)
}
\arguments{
\item{smr01}{Input tibble for admissions, see details.}

\item{smr01_minus5}{Input tibble for admissions going back five years, see
details.}

\item{morbs}{Input tibble for the charlson index for comorbidities lookup.}
}
\description{
Creates the remainder of necessary variables to be used in the
HSMR model.
}
\details{
\code{smr_pmorbs} expects a \code{tibble} of data extracted from
SMR01 that has already been through \code{smr_wrangling}.
It also expects a \code{tibble} of data extracted from SMR01 covering a
time-period that begins five years prior to that of the data in @param smr01.
This is so that the function is able to calculate the Charlson Index for
Comorbidities weighting for the previous five years.
It also expects a \code{tibble} for the Charlson Index lookups.
}
