% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr.R
\name{yr}
\alias{yr}
\title{Assign a a year to a date}
\usage{
yr(end_date)
}
\arguments{
\item{end_date}{The cut-off date for data to be included in the current HSMR
publication, supplied with \code{Date} class. Must be the final day of
either March, June, September or December.}
}
\description{
The \code{yr} functions take the end date of the extract period for the
Hospital Standardised Mortality Ratios Publication and returns the latest
year in text form.
}
\examples{
yr(end_date = lubridate::dmy(31032019))
}
