% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_hbcodes.R
\name{change_hbcodes}
\alias{change_hbcodes}
\title{Change health boards codes from 2019 version to the 2014 one.}
\usage{
change_hbcodes(dataset, version_to)
}
\arguments{
\item{dataset}{The object that you want to save.}

\item{version_to}{If you want to go 14 to 19 or 19 to 14.}
}
\description{
# Tableau uses 2014 codes, but code produces 2019. Changing from
2019 to 2014 for the TDE files
}
