% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeness.R
\name{completeness}
\alias{completeness}
\title{SMR01 completeness}
\usage{
completeness(quarter = c("previous", "current"), level = c("board",
  "scotland"), first_day)
}
\arguments{
\item{quarter}{A \code{character} string specifying the quarter for which
completeness data should be returned. Valid options are `previous` and
`current`.}

\item{level}{A \code{character} string specifying the geographic level for
which completeness data should be returned. Valid options are `board` and
`scotland`.}

\item{first_day}{The first day of the latest quarter, supplied with
\code{Date} class. Must be the first day of either January, April, July or
October. Only one value of \code{first_day} is valid at any one time. When
the spreadsheet is updated with a new quarter's data, the previously valid
value of \code{first_day} will now return an error, and a new value (the
previous value + 3 months) should be supplied instead.}
}
\value{
When \code{level} is set equal to \code{board}, \code{completeness}
returns a sentence listing the boards (if any) which had an SMR01
completeness level of less than 95 percent for the specified \code{quarter}.
When \code{level} is set equal to \code{scotland}, \code{completeness}
returns Scotland's SMR01 completeness level for the specified \code{quarter}
as a percentage.
}
\description{
\code{completeness} takes the first day of the latest quarter
of data which are included in the current HSMR publication. It extracts
completeness information for the SMR01 dataset from the
\href{https://www.isdscotland.org/Products-and-Services/Data-Support-and-Monitoring/SMR-Completeness/_docs/SMR_Estimates.xlsx}{SMR Completeness Estimates spreadsheet}
published on the
\href{https://www.isdscotland.org/}{ISD website}. It extracts the
completeness percentages for the two most recent quarters for all health
boards and Scotland.
}
\details{
The \code{completeness} function assumes that the list of health
board names (including Scotland) is contained in cells B31:B47 in the
spreadsheet, where B29 is `NHS Board` and B47 is `All NHS Boards`. Should
this change, the function should be edited accordingly.

The URL of the spreadsheet is occasionally modified. Should this happen, the
function should again be edited accordingly.

If the ISD website is down, the function will return an error saying it has
timed out. The package will also fail to build when the website is down, and
the unit tests will fail to run, as access to the spreadsheet is required.
Nothing can be done in this instance other than waiting for the website to
return to normal.
}
