% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_sizes.R
\name{file_sizes}
\alias{file_sizes}
\title{Extract file sizes for HSMR Excel tables}
\usage{
file_sizes(filepath = here::here("data", "output"))
}
\arguments{
\item{filepath}{A \code{character} string containing the filepath to the
location of the HSMR Excel tables. Errors if not provided with a valid
filepath. Defaults to \code{here::here("data", "output")}.}
}
\value{
A \code{character} vector.
}
\description{
\code{file_sizes} takes the filepath to the location of the
HSMR Excel tables. It returns a \code{character} vector containing the size
of each file.
}
\details{
There are three HSMR Excel tables: tables 1, 2 and 3. If fewer or
more Excel tables which follow the same naming conventions as tables 1, 2
and 3 are found in the filepath, the function will return an error.

File sizes are returned in kilobytes (KB) or megabytes (MB). A kilobyte is
taken to be 1,024 bytes, and a megabyte to be 1,024 kilobytes.
}
\examples{
file_sizes()
file_sizes(getwd())[1]

}
