% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnel_text.R
\name{funnel_text}
\alias{funnel_text}
\title{Create text for Funnel Plot outliers}
\usage{
funnel_text(smr_data)
}
\arguments{
\item{quarter}{Should the final month of the current or next quarter be
provided?}
}
\description{
test
}
\examples{
qtr(first_day = lubridate::dmy(01012018), format = "short")
qtr(first_day = lubridate::dmy(01102018), format = "long")

qtr_end(first_day = lubridate::dmy(01072018), quarter = "current")
qtr_end(first_day = lubridate::dmy(01042018), quarter = "next")

qtr_prev(first_day = lubridate::dmy(01012018))

}
