% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub_date.R
\name{pub_date}
\alias{pub_date}
\title{Obtain HSMR publication date}
\usage{
pub_date(end_date, pub = c("previous", "current", "next"))
}
\arguments{
\item{end_date}{The cut-off date for data to be included in the current HSMR
publication.}

\item{pub}{A character string specifying the publication of interest. Valid
options are `previous`, `current` and `next`.}
}
\description{
\code{pub_date} takes the final date for which SMR data are
included in the current publication. It uses this cut-off date to calculate
the publication dates of the previous, current and next HSMR publications.
}
\details{
\code{pub_date} accepts a single value of \code{Date} class only.
The only dates it accepts are the last day of the month in March, June,
September and December, as these are the only cut-off dates used in the HSMR
publication.
}
\examples{
extract_end_date <- lubridate::dmy(30092018)
pub_date(end_date = extract_end_date, pub = "previous")
pub_date(end_date = extract_end_date, pub = "current")
pub_date(end_date = extract_end_date, pub = "next")

}
