% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smr_data.R
\name{smr_data}
\alias{smr_data}
\title{Produce HSMR dataset}
\usage{
smr_data(smr01, index)
}
\arguments{
\item{smr01}{Input tibble for admissions, see details.}

\item{index}{To define whether data produced are to be quarterly or annual.}
}
\value{
If the class is not initiated correctly, nothing is returned.
}
\description{
Returns the minimal tidy data set for the HSMR dataset.
}
\details{
\code{smr_data} expects a \code{tibble} with data extracted
from SMR01 which has already been run through \code{smr_wrangling} and then
\code{smr_pmorbs} and finally \code{smr_model}. The data MUST be run through
these three functions first as \code{smr_data} is the final step of a
four-step process.
}
