% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smr_functions.R
\name{smr_wrangling}
\alias{smr_wrangling}
\title{Clean up logistic regression model}
\usage{
smr_wrangling(smr01, gro, pdiags, postcode, morbs)
}
\arguments{
\item{cm}{Input \code{glm}.}
}
\description{
Strips logistic regression model object of extraneous data that
are not required in order to calculate the probability of death within
30 days of admission. This is a necessary step as the object will likely be
quite large and take up significant amounts of memory when producing the
probabilities.
}
\details{
\code{clean_model} expects an object of class \code{"glm"}.
}
\examples{


}
