% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smr_wrangling.R
\name{smr_wrangling}
\alias{smr_wrangling}
\title{Initial Wrangling for HSMR data}
\usage{
smr_wrangling(smr01, gro, pdiags, postcode, morbs)
}
\arguments{
\item{smr01}{Input tibble for admissions, see details.}

\item{gro}{Input tibble for deaths, see details.}

\item{pdiags}{Input tibble for primary diagnosis groupings lookup.}

\item{postcode}{Input tibble for deprivation lookup.}

\item{morbs}{Input tibble for the charlson index for comorbidities lookup.}
}
\description{
Creates the majority of necessary variables to be used in the
HSMR model.
}
\details{
\code{smr_wrangling} expects a \code{tibble} of data extracted from
SMR01. It also expects \code{tibbles} of data extracted from the GRO deaths
database and lookups for primary diagnosis, postcode(deprivation) and
morbidities (comorbs_sum). This is the first step in a four-step process.
The next steps are \code{smr_pmorbs} then \code{smr_model} and finally
\code{smr_data}.
}
